requirements.txt
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
appdirs==1.4.4
argcomplete==3.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autocommand==2.2.2
Babel==2.14.0
bcc==0.30.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
Bottleneck==1.3.8
Brlapi==0.8.5
Brotli==1.1.0
certifi==2024.6.2
chardet==5.2.0
click==8.1.7
cloud-init==24.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
contourpy==1.2.1
cryptography==42.0.5
cssselect==1.2.0
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.9.0
distro-info==1.9
et-xmlfile==1.0.1
Flask==3.0.3
fonttools==4.46.0
frozenlist==1.4.1
fs==2.4.16
gyp-next==0.16.2
html5lib==1.1
httplib2==0.22.0
idna==3.6
importlib_metadata==8.2.0
inflect==7.3.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.0
jaraco.functools==4.0.2
jeepney==0.8.0
Jinja2==3.1.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.19.2
jsonschema-specifications==2023.12.1
keyring==25.3.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==2.0.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
llvmlite==0.42.0
louis==3.30.0
lxml==5.3.0
lz4==4.0.2+dfsg
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
netaddr==0.10.1
netifaces==0.11.0
numba==0.59.1
numexpr==2.10.1
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.2
olefile==0.46
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2+dfsg
pefile==2023.2.7
pexpect==4.9.0
pillow==10.4.0
pipx==1.6.0
platformdirs==4.2.2
proton-core==0.4.0
proton-keyring-linux==0.2.0
proton-vpn-api-core==0.39.0
proton-vpn-gtk-app==4.8.2
proton-vpn-network-manager==0.10.2
psutil==5.9.8
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyasyncore==1.0.2
pycairo==1.26.1
pycares==4.4.0
pycryptodomex==3.20.0
pycups==2.0.4
Pygments==2.18.0
PyGObject==3.48.2
pyinotify==0.9.6
PyJWT==2.7.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pyserial==3.5
python-apt==2.9.0+ubuntu1
python-dateutil==2.9.0
python-debian==0.1.49+ubuntu3
python-gnupg==0.5.2
python-magic==0.4.27
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.2
QuantLib==1.35
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
scipy==1.13.1
SecretStorage==3.3.3
sentry-sdk==1.40.4
simplejson==3.19.2
six==1.16.0
soupsieve==2.6
speedtest-cli==2.1.3
sympy==1.13.1
systemd-python==235
tables==3.10.0
typeguard==4.3.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.16.0
ufw==0.36.2
unattended-upgrades==0.1
unicodedata2==15.1.0
urllib3==2.0.7
userpath==1.9.1
vboxapi==1
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==3.0.3
wheel==0.44.0
xdg==5
xkit==0.0.0
yara-python==4.5.1
yarl==1.9.4
zipp==3.20.0
